pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "nadf"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "adversarial-robustness-toolbox>=1.18.0",
    "jupyter>=1.1.1",
    "matplotlib>=3.9.4",
    "numpy>=2.0.2",
    "pandas>=2.3.3",
    "python-dotenv>=1.1.1",
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "tqdm>=4.66.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src/nadf"]

[dependency-groups]
dev = [
    "black>=25.9.0",
    "pytest>=8.4.2",
    "ruff>=0.13.3",
]

[tool.ruff]
# Exclude common directories
exclude = [
    ".git",
    ".venv",
    "__pycache__",
    "build",
    "dist",
    "*.egg-info",
]
line-length = 120  # Increased from 100 to accommodate longer lines
target-version = "py39"

[tool.ruff.lint]
# Enable common linting rules
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # pyflakes
    "I",      # isort (import sorting)
    "N",      # pep8-naming
    "UP",     # pyupgrade
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "SIM",    # flake8-simplify
]
ignore = []

[tool.ruff.lint.isort]
known-first-party = ["nadf"]
